+++ src/components/editor/index.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script lang="ts">
-import { toRefs, reactive, onMounted } from 'vue';
+import { toRefs, reactive, onMounted, watch } from 'vue';
 import wangeditor from 'wangeditor';
 export default {
 	name: 'wngEditor',
@@ -20,6 +20,11 @@ export default {
 			type: Boolean,
 			default: () => false,
 		},
+		// 内容框默认 placeholder
+		placeholder: {
+			type: String,
+			default: () => '请输入内容',
+		},
 		// 双向绑定
 		// 双向绑定值，字段名为固定，改了之后将不生效
 		// 参考：https://v3.cn.vuejs.org/guide/migration/v-model.html#%E8%BF%81%E7%A7%BB%E7%AD%96%E7%95%A5
@@ -32,8 +37,9 @@ export default {
 		// 初始化富文本
 		// https://doc.wangeditor.com/
 		const initWangeditor = () => {
-			state.editor = new wangeditor('#wangeditor');
-			state.editor.config.placeholder = '请输入内容';
+			state.editor = new wangeditor(`#${props.id}`);
+			state.editor.config.zIndex = 1;
+			state.editor.config.placeholder = props.placeholder;
 			state.editor.config.uploadImgShowBase64 = true;
 			state.editor.config.showLinkImg = false;
 			onWangeditorChange();
@@ -51,6 +57,14 @@ export default {
 		onMounted(() => {
 			initWangeditor();
 		});
+		// 监听双向绑定值的改变
+		// https://gitee.com/lyt-top/vue-next-admin/issues/I4LM7I
+		watch(
+			() => props.modelValue,
+			(value) => {
+				state.editor.txt.html(value);
+			}
+		);
 		return {
 			...toRefs(state),
 		};