+++ src/utils/other.ts
@@ -1,6 +1,6 @@
 import { nextTick } from 'vue';
 import type { App } from 'vue';
-import * as svg from '@element-plus/icons';
+import * as svg from '@element-plus/icons-vue';
 import router from '/@/router/index';
 import { store } from '/@/store/index';
 import { i18n } from '/@/i18n/index';
@@ -25,16 +25,14 @@ export function elSvg(app: App) {
  * @method const title = useTitle(); ==> title()
  */
 export function useTitle() {
-	return () => {
-		nextTick(() => {
-			let webTitle = '';
-			let globalTitle: string = store.state.themeConfig.themeConfig.globalTitle;
-			router.currentRoute.value.path === '/login'
-				? (webTitle = router.currentRoute.value.meta.title as any)
-				: (webTitle = i18n.global.t(router.currentRoute.value.meta.title as any));
-			document.title = `${webTitle} - ${globalTitle}` || globalTitle;
-		});
-	};
+	nextTick(() => {
+		let webTitle = '';
+		let globalTitle: string = store.state.themeConfig.themeConfig.globalTitle;
+		router.currentRoute.value.path === '/login'
+			? (webTitle = router.currentRoute.value.meta.title as any)
+			: (webTitle = i18n.global.t(router.currentRoute.value.meta.title as any));
+		document.title = `${webTitle} - ${globalTitle}` || globalTitle;
+	});
 }
 
 /**